+++ app/src/main/java/com/giussepr/mubi/data/database/MubiDatabase.kt
@@ -11,8 +11,8 @@ import com.giussepr.mubi.data.database.dao.FavoriteTvShowDao
 import com.giussepr.mubi.data.database.dao.TopRatedTvShowDao
 import com.giussepr.mubi.data.database.dao.TopRatedTvShowRemoteKeyDao
 import com.giussepr.mubi.data.database.entity.FavoriteTvShowEntity
-import com.giussepr.mubi.data.database.entity.TopRatedTvShowEntity
-import com.giussepr.mubi.data.database.entity.TopRatedTvShowRemoteKey
+import com.giussepr.mubi.data.database.entity.topratedtvshow.TopRatedTvShowEntity
+import com.giussepr.mubi.data.database.entity.topratedtvshow.TopRatedTvShowRemoteKey
 import com.giussepr.mubi.data.database.migrations.Migration1
 import com.giussepr.mubi.data.database.migrations.Migration2
 import com.giussepr.mubi.data.database.migrations.Migration3
